+++ crates/symbolicator/src/cli.rs
@@ -35,7 +35,11 @@ enum Command {
 
     /// Clean local caches.
     #[command(name = "cleanup")]
-    Cleanup,
+    Cleanup {
+        /// Only simulate the cleanup without deleting any files.
+        #[arg(long)]
+        dry_run: bool,
+    },
 }
 
 /// Command line interface parser.
@@ -88,10 +92,10 @@ pub fn execute() -> Result<()> {
         session_mode: sentry::SessionMode::Request,
         auto_session_tracking: false,
         traces_sampler: Some(Arc::new(move |ctx| {
-            if Some(true) == ctx.sampled() {
+            if Some(true) == ctx.sampled() && config.propagate_traces {
                 1.0
             } else if ctx.operation() != "http.server" {
-                config.transaction_sample_rate
+                config.traces_sample_rate
             } else {
                 0.0
             }
@@ -153,7 +157,9 @@ pub fn execute() -> Result<()> {
 
     match cli.command {
         Command::Run => server::run(config).context("failed to start the server")?,
-        Command::Cleanup => caching::cleanup(config).context("failed to clean up caches")?,
+        Command::Cleanup { dry_run } => {
+            caching::cleanup(config, dry_run).context("failed to clean up caches")?
+        }
     }
 
     Ok(())